### 2020-02-10 #2092 @whedon
[PRE REVIEW]: Visions: An Open-Source Library for Semantic Data
**Submitting author:** @sbrugman (<a href="http://orcid.org/0000-0001-9866-7767">Simon Brugman</a>)
**Repository:** <a href="https://github.com/dylan-profiler/visions" target ="_blank">https://github.com/dylan-profiler/visions</a>
**Version:** v0.2.2
