### 2020-02-10 #2092 @danielskatz
👋 @openjournals/dev - Can whedon compile papers with figures that are remote?  If so, any idea what the problem is here?
