+++ paper/paper.bib
@@ -11,13 +11,16 @@
 
 # Numpy
 @article{van2011numpy,
-  title={The NumPy array: a structure for efficient numerical computation},
-  author={Van Der Walt, Stefan and Colbert, S Chris and Varoquaux, Gael},
-  journal={Computing in Science \& Engineering},
+  journal={Computing in Science Engineering},
+  title={The NumPy Array: A Structure for Efficient Numerical Computation},
+  year={2011},
   volume={13},
   number={2},
-  pages={22},
-  year={2011},
+  pages={22-30},
+  keywords={data structures;high level languages;mathematics computing;numerical analysis;numerical computation;numpy array;numerical data;high level language;Python programming language;Arrays;Numerical analysis;Performance evaluation;Computational efficiency;Finite element methods;Vector quantization;Resource management;Python;NumPy;scientific programming;numerical computations;programming libraries},
+  doi={10.1109/MCSE.2011.37},
+  ISSN={1558-366X},
+  month={March},
   publisher={IEEE Computer Society}
 }
 
