### 2020-02-10 #2092 @sbrugman
@danielskatz The figure pointed to the branch we used for development. Moreover, the numpy reference was missing it's DOI. I've fixed both issues.
