### 2020-02-10 #2092 @danielskatz
You also might want to take the time to fix case issues in the references - use {}s as needed to protect case, such as {P}ython.
