+++ paper/paper.bib
@@ -1,6 +1,6 @@
 # Pandas
 @inproceedings{mckinney2010data,
-  title={Data structures for statistical computing in python},
+  title={{D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython},
   author={McKinney, Wes and others},
   booktitle={Proceedings of the 9th Python in Science Conference},
   volume={445},
@@ -12,7 +12,8 @@
 # Numpy
 @article{van2011numpy,
   journal={Computing in Science Engineering},
-  title={The NumPy Array: A Structure for Efficient Numerical Computation},
+  title={{T}he {N}um{P}y {A}rray: {A} {S}tructure for {E}fficient {N}umerical {C}omputation},
+  author={Walt, St{\'e}fan van der and Colbert, S Chris and Varoquaux, Gael},
   year={2011},
   volume={13},
   number={2},
@@ -27,7 +28,7 @@
 # Networkx
 @InProceedings{SciPyProceedings_11,
   author =       {Aric A. Hagberg and Daniel A. Schult and Pieter J. Swart},
-  title =        {Exploring Network Structure, Dynamics, and Function using NetworkX},
+  title =        {{E}xploring {N}etwork {S}tructure, {D}ynamics, and {F}unction using {N}etworkX},
   booktitle =   {Proceedings of the 7th Python in Science Conference},
   pages =     {11 - 15},
   address = {Pasadena, CA USA},
@@ -77,7 +78,7 @@
 }
 
 @article{stroustrup1982classes,
-  title={Classes: an abstract data type facility for the C language},
+  title={{C}lasses: an abstract data type facility for the {C} language},
   author={Stroustrup, Bjarne},
   journal={ACM Sigplan Notices},
   volume={17},
@@ -88,7 +89,7 @@
 }
 
 @article{stonebraker2018data,
-  title={Data Integration: The Current Status and the Way Forward.},
+  title={{D}ata {I}ntegration: {T}he {C}urrent {S}tatus and the {W}ay {F}orward},
   author={Stonebraker, Michael and Ilyas, Ihab F},
   journal={IEEE Data Eng. Bull.},
   volume={41},
@@ -99,7 +100,7 @@
 
 @misc{pandasprofiling,
   author = "Brugman, Simon",
-  title = "{pandas-profiling}: exploratory data analysis reports in Python",
+  title = "{pandas-profiling}: exploratory data analysis reports in {P}ython",
   year = "2020",
   howpublished = "https://github.com/pandas-profiling/pandas-profiling",
   note = "Version: 2.4.0, Accessed: 01/02/2020"
