### 2020-02-11 #2092 @arfon
@sbrugman - please change links to images to make them full URLs, e.g. change `![Visions' Standard Typeset](images/typeset_standard.png)` to `https://raw.githubusercontent.com/dylan-profiler/visions/master/paper/images/typeset_standard.png`

Also, please cite your BibTeX entries using the rMarkdown syntax. You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)
