### 2020-03-03 c7d8c5d
Evolve types (i.e. create new type with additional relation)
Fix ordinal detection bug

Several fixes in `lib/relations`

Update `setup.py` to read from `requirements.txt`

Documentation:
- Docstrings functional API
- Extending documentation updated
- Document functional API for diffing type changes
- Update citation year
