+++ src/visions/lib/relations/integer_to_count.py
@@ -1,8 +1,7 @@
 import numpy as np
 import pandas as pd
-from visions import visions_integer, visions_count
+from visions import visions_integer
 
-from visions.core.model import TypeRelation
 from visions.core.model.relations import InferenceRelation
 
 
@@ -11,10 +10,10 @@ def is_unsigned_int(series: pd.Series):
     return series.ge(0).all()
 
 
-def integer_to_count():
+def integer_to_count(cls) -> InferenceRelation:
     return InferenceRelation(
         relationship=is_unsigned_int,
         transformer=lambda s: s.astype(np.uint64),
-        type=visions_integer,
-        related_type=visions_count,
+        type=cls,
+        related_type=visions_integer,
     )
