### 2020-03-04 #2092 @sbrugman
@mstimberg @macrocosme @pandaconstantin @rchg @fccoelho Hello! Would you be interested in reviewing this submission?

@mjsottile People might be put off by thinking they are not familiar enough with the topic, as the title does not include downstream applications. How `visions` relates to e.g. machine learning, data compression is described here: https://dylan-profiler.github.io/visions/visions/applications/applications.html. Moreover, a background in functional programming could also provide a good intuition of the contribution of this work.
