### 2020-03-04 #2145 @whedon
[REVIEW]: Visions: An Open-Source Library for Semantic Data
**Submitting author:** @sbrugman (<a href="http://orcid.org/0000-0001-9866-7767">Simon Brugman</a>)
**Repository:** <a href="https://github.com/dylan-profiler/visions" target ="_blank">https://github.com/dylan-profiler/visions</a>
**Version:** v0.3.0
