### 2020-03-09 #2145 @1313e
I am currently reviewing this, and I am going to assume that I should review v0.2.3, as that version has the bugfix for the `attr.s` decorator.
Can this be confirmed and if so, updated above? @sbrugman @mjsottile
