### 2020-03-09 dylan-profiler/visions#21 @1313e
Outdated online documentation
(This is part of the review in openjournals/joss-reviews#2145)

The online documentation at the moment mentions that the version is v0.2.1, while the latest version on PyPI is v0.2.3.
Can the online documentation please be updated to the same version as the released package?
