### 2020-03-09 #2145 @mjsottile
@1313e I would review the version with the bug fix.  When the process is finished, the version number for the submission will be set to whatever version is archived - which I assume will reflect the bug fix you mention, as well as any other bugs and changes that arise during the review process.
