### 2020-03-09 dylan-profiler/visions#22 @1313e
API and Usage
(This is part of the review in openjournals/joss-reviews#2145)

Hi @sbrugman,

I am currently going through the package and I found it a very interesting project.
The type inference that already exists in built-in Python is frequently not enough, and I often find myself writing my own functions for it on a case-by-case basis.
So, it is nice to see that this is being done.

However, I do find myself having quite a bit of trouble using the package effectively.
Funnily enough, this is mostly caused by the (in my opinion) confusing naming schemes and the structure of the namespace.
It may require a bit of effort to solve (not to mention that it might create incompatibilities with previous versions), but I believe that it will greatly improve the user experience when fixed.

So, the main problem in my opinion is that almost all definitions are stored in their own subpackages in the `visions.core` subpackage, often separated as well over an `implementations` and a `model` subpackage.
This means that in order to access a definition, let's say `VisionsTypeset`, I have to import this from `visions.core.model.typeset`, while pre-defined typesets I have to import from `visions.core.implementations.typesets`.
In my opinion, it is incredibly confusing that these are stored in different subpackages/submodules, as they are related to the same thing, namely typesets, and I would expect to find all of these definitions in a `visions.typesets` subpackage.
Preferably, all definitions the average user would use, should be available either at root (`visions`) or a single level deep (`visions.xxx`).

I also noticed that almost all definitions have the word `visions` in their name.
I get the feeling that the reason for this is to avoid namespace clashes when someone uses a wildcard import (`from visions import *`).
However, as wildcard imports are heavily discouraged in the community, this leads to the user writing the word `visions` at least twice for every definition used (for example, using the `visions_integer` type requires me to write `visions.visions_integer`, which could be simplified to `visions.integer` or even `visions.int`).

Finally, I am not entirely sure if this has to do with the online documentation being outdated as mentioned in #21, but according to the example [here](https://dylan-profiler.github.io/visions/visions/getting_started/usage/typesets.html#detection), a `visions_standard_set` object has a method called `detect_type_series`.
In v0.2.3, this object neither has a method called `detect_type_series` nor `type_detect_series` (the name that the stand-alone function has in `visions.core.functional`), but instead it is called `detect_series_type`.
If possible, could you check and make sure that the methods and stand-alone functions use consistent naming schemes?

Please let me know if you have any questions.
