### 2020-03-09 dylan-profiler/visions#21 @sbrugman
Hi Ellert,

Thank you for reporting this issue. We've updated the docs in #19 and will update the package together with the first round of joss feedback.
