### 2020-03-09 dylan-profiler/visions#22 @sbrugman
Hi @1313e,

Great to hear you like our project.

Thanks for your extensive feedback. I agree that we can improve the user interface. The rationale behind distinguishing between the framework (model) and predefined types was the anticipation of additional typesets (statistics, finance) that are possibly incompatible with current definitions. In retrospect, we can hugely simplify this. Your point about class naming also makes sense.

The documentation should be updated. The expected behaviour is to allow both OO and functional styles: `typeset.detect_series_type(series)` and `detect_series_type(typeset, series)`.
