### 2020-03-09 dylan-profiler/visions#23 @fccoelho
Expand Example applications (Joss review)
Hi,
this is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/2145).

After reading the official docs, I found that the [Application  section](https://dylan-profiler.github.io/visions/visions/applications/applications.html) is rather underdeveloped.
I think that providing complete examples here, would really help new users to better grasp the utility of this library.
Apart from `data sumarization` which include a code  block with a brief example, the remaining applications have no tangible example available.
So please consider adding meaningful examples, with code, in this section.
