### 2020-03-09 dylan-profiler/visions#23 @sbrugman
The purpose of the application section is to provide inspiration for potential applications and to understand the utility for more pratically oriented users. This can be articualted more clearly. Thank you.
