### 2020-03-09 6f6f127
Terminology paper and documentation (#24).
