+++ docs/_sources/visions/api/_autosummary/visions.core.functional.type_inference_report_frame.rst.txt
@@ -1,0 +1,6 @@
+visions.core.functional.type\_inference\_report\_frame
+======================================================
+
+.. currentmodule:: visions.core.functional
+
+.. autofunction:: type_inference_report_frame
