+++ docs/_sources/visions/api/_autosummary/visions.core.implementations.types.visions_uuid.rst.txt
@@ -15,7 +15,7 @@ visions.core.implementations.types.visions\_uuid
    
       ~visions_uuid.__init__
       ~visions_uuid.contains_op
-      ~visions_uuid.extend_relations
+      ~visions_uuid.evolve_relations
       ~visions_uuid.get_relations
    
    
