+++ docsrc/source/visions/applications/validation.rst
@@ -7,9 +7,9 @@ For example, we might expect a field to always be unique (identifier) or a numbe
 Tight coupling causes the same problem as discussed above because the user often designs validation rules that are grounded in the meaning of the data.
 
 A variety of software packages have been developed to facilitate this workflow [#f1]_, [#f2]_, [#f3]_, [#f4]_, [#f5]_, [#f6]_.
-These packages can benefit from decoupling semantic and physical types.
+These packages can benefit from decoupling semantic and machine type representations.
 
-The part of data validation that depends on physical types, the validation methods, should have to focus of such a package.
+The part of data validation that depends on machine types, the validation methods, should have to focus of such a package.
 For example, the software could have helper functions to assert if a float is approximately the reference value (see `pytest.approx` in [pytest]_).
 The other part, defining and checking properties of the data depends on semantic types.
 For example, the International Standard Book Number (ISBN) has a `check digit <https://en.wikipedia.org/wiki/Check_digit>`_, which only makes sense to validate when a stored number is representing the ISBN.
