### 2020-03-09 3910fea
(wip) Restructuring package
According to #22, the API of `visions` made it confusing for new users to effectively use the package. Related classes were not in the same namespace/subpackage, despite this would be a natural expectation.
