+++ src/visions/typesets/typeset.py
@@ -1,0 +1,427 @@
+import warnings
+from typing import Type, Tuple, List, Dict, Iterable, Optional
+
+import pandas as pd
+import networkx as nx
+
+from visions.types.type import VisionsBaseType
+from visions.types.visions_generic import visions_generic
+
+
+def build_graph(nodes: set) -> Tuple[nx.DiGraph, nx.DiGraph]:
+    """Constructs a traversable relation graph between visions types
+    Builds a type relation graph from a collection of root and derivative nodes. Usually
+    root nodes correspond to the baseline numpy types found in pandas while derivative
+    nodes correspond to subtypes with a defined relation.
+
+    Args:
+        nodes:  A list of vision_types considered at the root of the relations graph.
+
+    Returns:
+        A directed graph of type relations for the provided nodes.
+    """
+
+    style_map = {True: "dashed", False: "solid"}
+    relation_graph = nx.DiGraph()
+    relation_graph.add_nodes_from(nodes)
+
+    noninferential_edges = []
+
+    for node in nodes:
+        for relation in node.get_relations():
+            if relation.related_type not in nodes:
+                warnings.warn(
+                    "Provided relations included mapping from {related_type} to {own_type} but {related_type} was not included in the provided list of nodes".format(
+                        related_type=relation.related_type, own_type=relation.type
+                    )
+                )
+
+            relation_graph.add_edge(
+                relation.related_type,
+                relation.type,
+                relationship=relation,
+                style=style_map[relation.inferential],
+            )
+
+            if not relation.inferential:
+                noninferential_edges.append((relation.related_type, relation.type))
+
+    check_graph_constraints(relation_graph)
+    return relation_graph, relation_graph.edge_subgraph(noninferential_edges)
+
+
+def check_graph_constraints(relation_graph: nx.DiGraph) -> None:
+    """Validates a relation_graph is appropriately constructed
+
+    Args:
+        relation_graph: A directed graph representing the set of relations between type nodes.
+
+    """
+    check_isolates(relation_graph)
+    check_cycles(relation_graph)
+
+
+def check_isolates(graph: nx.DiGraph) -> None:
+    """Check for orphaned nodes.
+
+    Args:
+        graph: the graph to check
+
+    """
+    nodes = set(graph.nodes)
+    isolates = list(set(nx.isolates(graph)) - {visions_generic})  # root can be isolate
+    graph.remove_nodes_from(isolates)
+    orphaned_nodes = nodes - set(graph.nodes)
+    if orphaned_nodes:
+        warnings.warn(
+            "{orphaned_nodes} were isolates in the type relation map and consequently\
+                      orphaned. Please add some mapping to the orphaned nodes.".format(
+                orphaned_nodes=orphaned_nodes
+            )
+        )
+
+
+def check_cycles(graph: nx.DiGraph) -> None:
+    """Check for cycles and warn if one is found
+
+    Args:
+        graph: the graph to check
+
+    """
+    cycles = list(nx.simple_cycles(graph))
+    if len(cycles) > 0:
+        warnings.warn(
+            "Cyclical relations between types {cycles} detected".format(cycles=cycles)
+        )
+
+
+def traverse_graph(
+    series: pd.Series, graph: nx.DiGraph, node: Type[VisionsBaseType] = visions_generic
+) -> Type[VisionsBaseType]:
+    """Depth First Search traversal. There should be at most one successor that contains the series.
+
+    Args:
+        series: the Series to check
+        graph: the Graph to traverse
+        node: the current node
+
+    Returns:
+        The most uniquely specified node matching the series.
+    """
+    for vision_type in graph.successors(node):
+        if series in vision_type:
+            return traverse_graph(series, graph, vision_type)
+
+    return node
+
+
[…]
