### 2020-03-09 dylan-profiler/visions#22 @1313e
You could invert the logic a bit, by for example for the typesets, make a subpackage called `visions.typesets.core.implementations` and `visions.typesets.core.model`.
This way, you can keep the structure you already have, but you can make sure that everything is available under `visions.typesets`.
Then, for the user, it does not really matter what the underlying structure is, as long as it is under `visions.typesets`.
