### 2020-03-10 891436d
(wip) Restructuring package
- Naming convention for concepts are not CamelCase without `visions_` prefix.
- Added a documentation section on importing the package.
- Add placeholders for type namespaces (e.g. statistics, finance)
- Remove legacy code dtypes
