+++ src/visions/types/__init__.py
@@ -1,23 +1,23 @@
 from visions.types.type import VisionsBaseType
 
-from visions.types.visions_generic import visions_generic
-from visions.types.visions_string import visions_string
-from visions.types.visions_bool import visions_bool
-from visions.types.visions_categorical import visions_categorical
-from visions.types.visions_complex import visions_complex
-from visions.types.visions_count import visions_count
-from visions.types.visions_date import visions_date
-from visions.types.visions_datetime import visions_datetime
-from visions.types.visions_existing_path import visions_existing_path
-from visions.types.visions_float import visions_float
-from visions.types.visions_geometry import visions_geometry
-from visions.types.visions_image_path import visions_image_path
-from visions.types.visions_integer import visions_integer
-from visions.types.visions_ip import visions_ip
-from visions.types.visions_object import visions_object
-from visions.types.visions_ordinal import visions_ordinal
-from visions.types.visions_path import visions_path
-from visions.types.visions_time import visions_time
-from visions.types.visions_timedelta import visions_timedelta
-from visions.types.visions_url import visions_url
-from visions.types.visions_uuid import visions_uuid
+from visions.types.generic import Generic
+from visions.types.string import String
+from visions.types.boolean import Boolean
+from visions.types.categorical import Categorical
+from visions.types.complex import Complex
+from visions.types.count import Count
+from visions.types.date import Date
+from visions.types.date_time import DateTime
+from visions.types.existing_path import ExistingPath
+from visions.types.float import Float
+from visions.types.geometry import Geometry
+from visions.types.image_path import ImagePath
+from visions.types.integer import Integer
+from visions.types.ip_address import IPAddress
+from visions.types.object import Object
+from visions.types.ordinal import Ordinal
+from visions.types.path import Path
+from visions.types.time import Time
+from visions.types.time_delta import TimeDelta
+from visions.types.url import URL
+from visions.types.uuid import UUID
