+++ tests/test_type_convert.py
@@ -2,12 +2,12 @@ import sys
 
 import pytest
 
-from visions.typesets import visions_complete_set
+from visions.typesets import CompleteSet
 from tests.series import get_series, get_convert_map
 
 
 def all_relations_tested(series_map):
-    typeset = visions_complete_set()
+    typeset = CompleteSet()
 
     # Convert data structure for mapping
     series_map_lookup = {}
@@ -98,7 +98,7 @@ def test_relations(source_type, relation_type, series):
 
 @pytest.mark.run(order=10)
 def test_consistency(series):
-    typeset = visions_complete_set()
+    typeset = CompleteSet()
 
     if (
         series.name
@@ -129,7 +129,7 @@ def test_consistency(series):
 def test_side_effects(series):
     reference = series.copy()
 
-    typeset = visions_complete_set()
+    typeset = CompleteSet()
     typeset.detect_series_type(series)
     typeset.infer_series_type(series)
 
@@ -145,7 +145,7 @@ def test_multiple_inference(series):
     Notes:
         Copy to prevent possible side effects only for testing.
     """
-    ts = visions_complete_set()
+    ts = CompleteSet()
 
     inferred_type = ts.infer_series_type(series)
 
