### 2020-03-10 c58b174
(wip) Practise what you preach (update `test_summarization.py` imports)
