### 2020-03-10 da5eb32
(wip) Update the image in the paper
