### 2020-03-10 d4c93ad
Rebased #19 on master without docs build
