### 2020-03-10 dylan-profiler/visions#22 @sbrugman
If I'm not mistaken, all points in this issue are addressed. The docs are here:
https://dylan-profiler.github.io/visions/visions/getting_started/usage/import.html

Thanks again!
I'll rerender the paper at the JOSS review when all PRs are merged.
