### 2020-03-10 dylan-profiler/visions#24 @sbrugman
The paper and docs are updated, we'll rerender when all PRs are merged.
