### 2020-03-10 df5dc3d
Clarify the purpose of the application section (#23).
