+++ docsrc/source/visions/applications/applications.rst
@@ -2,8 +2,9 @@
 Applications
 ============
 
-This section discusses possible downstream tasks where `visions` can contribute to simplifying the workflow.
-
-The potential applications are broad.
+This section discusses possible downstream tasks where `visions` could contribute to simplifying the workflow.
+Without a doubt, the potential applications are broader than we have currently described.
 We invite additional applications, if you'd like to share, via a pull request.
 You can find more info :doc:`on this page <../contributing/contributing>`
+
+It should be noted that the applications are meant as inspiration. Fully working solutions are out of the scope of this package, even though we would be excited to see them. For complete code examples, please have a look at the :doc:`examples page <../getting_started/examples>`
