+++ docsrc/source/visions/applications/applications.rst
@@ -3,8 +3,10 @@ Applications
 ============
 
 This section discusses possible downstream tasks where `visions` could contribute to simplifying the workflow.
-Without a doubt, the potential applications are broader than we have currently described.
-We invite additional applications, if you'd like to share, via a pull request.
-You can find more info :doc:`on this page <../contributing/contributing>`
+Although this is by no means an exhaustive list, let's explore a few obvious examples where `visions` can help contribute to simplifying downstream workflows.
+We invite additional applications if you'd like to share, via a pull request.
+You can find more information :doc:`on this page <../contributing/contributing>`.
 
-It should be noted that the applications are meant as inspiration. Fully working solutions are out of the scope of this package, even though we would be excited to see them. For complete code examples, please have a look at the :doc:`examples page <../getting_started/examples>`
+It should be noted that the applications are meant as inspiration.
+Fully working solutions are out of the scope of this package, even though we would be excited to see them.
+For complete code examples, please have a look at the :doc:`examples page <../getting_started/examples>`.
