### 2020-03-10 ee78787
More robust implementation for type evolution
