### 2020-03-10 dylan-profiler/visions#22 @1313e
Hi @sbrugman,

this looks really great now, exactly as I meant.
It makes it much easier to find your way around the package.

One side note, that you may want to incorporate (not necessary however): Currently, both the module and the class contained in that module show up in `visions.typesets`, `visions.types`, etc.
You can hide the modules by renaming them such that they start with an underscore, like renaming `visions.types.integer` to `visions.types._integer`.
As the only definition the user ever needs from this module is `Integer`, which is already imported into the `visions.types` subpackage, an average user will never need to access the module it came from.
So, you can opt into hiding it.

Keep in mind that this is a minor thing and absolutely not required.
It simply cleans up the namespace of the package a bit.

Besides that, excellent job on improving the package namespace!
