### 2020-03-11 e873419
adding relations property to BaseTypef
