### 2020-03-11 dylan-profiler/visions#22 @1313e
After looking at the docs again, I see that the last point that I mentioned has not been resolved yet.
Currently, the example [here](https://dylan-profiler.github.io/visions/visions/getting_started/usage/typesets.html#detection) still mentions the use of `detect_type_series`, whereas the `VisionsTypeset` class uses `detect_series_type`.
It is also still inconsistent with the function used from `visions.functional` (`type_detect_series`).
