### 2020-03-11 dylan-profiler/visions#30 @1313e
Improve community guidelines
(This is part of the review in openjournals/joss-reviews#2145)

Currently, the community guidelines [here](https://dylan-profiler.github.io/visions/visions/contributing/contributing.html) show what is expected if someone wants to contribute to the software (which is well done).
However, it currently does not mention how someone can seek support if they have issues with the software (like, by opening a GitHub issue).
Would be great if this could be added.
