### 2020-03-11 dylan-profiler/visions#31 @1313e
No module named 'shapely'
(This is part of the review in openjournals/joss-reviews#2145)

When I try to execute the example given [here](https://dylan-profiler.github.io/visions/visions/getting_started/examples/casting.html#casting-example), I get an error stating `No module named 'shapely'`.
I see that this is a dependency of `visions`, but is only listed in the [`requirements_test.txt`](https://github.com/dylan-profiler/visions/blob/master/requirements_test.txt).
You probably have to add this requirement to the `requirements.txt` as well.

PS: Currently, the requirements of the package are both listed in their own separate file and in the `setup.py` file.
To avoid confusion for yourself, it is probably better to only use either.
You can read in a requirements file and use it in the `setup.py` file by using:
```python
# Get the requirements list
with open('requirements.txt', 'r') as f:
    requirements = f.read().splitlines()
```

Keep in mind that it is possible to link different requirements files together.
For example, you can link `requirements.txt` and `requirements_dev.txt` together by adding the line `-r requirements.txt` to the top of `requirements_dev.txt`.
This means that installing the requirements of `requirements_dev.txt` will use both files.
This however won't work if you parse the file in a `setup.py` file.
In that case, you can simply read both files and append them together if necessary.
