### 2020-03-11 dylan-profiler/visions#32 @1313e
Pytest expected failures
(This is part of the review in openjournals/joss-reviews#2145)

When I run the tests, I see that there are many that are flagged as expected failures (actually, almost all of them are).
Although most of these tests are testing something that should not work from a package stand-point (therefore you expect them to fail), this is not what pytest's `xfail` marker should be used for.
The `xfail` marker is used for marking tests that are expected to fail, because they are testing something that is broken; not yet implemented; depends on a previous test to succeed that did not succeed; cannot succeed due to external influences (like running on a different OS that has a known issue); etc.
It is used for marking a test that is expected to not behave the way it should, therefore you expect it to fail.
A test should never be marked with `xfail` as a standard.

Instead, if you want to test if executing something in the package raises an error, which would be a test for correct exception handling, pytest's `pytest.raises` context manager should be used instead.
So, for example, if you want to test if a specific line raises a `ValueError`, you would write this as:
```python
def test_raises_ValueError():
    with pytest.raises(ValueError):
        this_raises_an_error()
```
The reason why you want to do this, is because pytest will mark this test as 'success' if a `ValueError` was actually raised, while it will mark it as 'failed' if it did not, which is exactly what you want.
On the other hand, a test marked as `xfail` will be marked as such regardless of it succeeding or failing, and depending on the settings, might not even be ran at all.
