### 2020-03-11 dylan-profiler/visions#22 @sbrugman
> Hi @sbrugman,
>
> this looks really great now, exactly as I meant.
> It makes it much easier to find your way around the package.
>
> One side note, that you may want to incorporate (not necessary however): Currently, both the module and the class contained in that module show up in `visions.typesets`, `visions.types`, etc.
> You can hide the modules by renaming them such that they start with an underscore, like renaming `visions.types.integer` to `visions.types._integer`.
> As the only definition the user ever needs from this module is `Integer`, which is already imported into the `visions.types` subpackage, an average user will never need to access the module it came from.
> So, you can opt into hiding it.
>
> Keep in mind that this is a minor thing and absolutely not required.
> It simply cleans up the namespace of the package a bit.
>
> Besides that, excellent job on improving the package namespace!

Certainly, there are modules that do not have to be accessible to the user. Moreover, there are certain modules that are problematic when imported in the user's namespace (e.g. `from visions.types import float` would overwrite the buildin float).

From my point of view, there is a simpler solution. We can specify `__all__` in `__init__.py` files: https://docs.python.org/3/tutorial/modules.html#importing-from-a-package. Surely, there is a tradeoff between renaming each file and keeping the `__all__` lists up to date.
