+++ src/visions/types/type.py
@@ -34,10 +34,11 @@ class VisionsBaseType(metaclass=VisionsBaseTypeMeta):
 
     Provides a common API for building custom visions data types.
     """
-    _relations : Optional[Sequence] = None
+
+    _relations: Optional[Sequence] = None
 
     def __init__(self):
-        raise ValueError('Types cannot be initialized')
+        raise ValueError("Types cannot be initialized")
 
     @classmethod
     @abstractmethod
@@ -58,7 +59,7 @@ class VisionsBaseType(metaclass=VisionsBaseTypeMeta):
         cls,
         type_name: str,
         relations_generator: Optional[Callable] = None,
-        replace: bool = False
+        replace: bool = False,
     ):
         """Make a copy of the type with the relations replaced by the relations return by `relations_func`.
 
@@ -70,16 +71,20 @@ class VisionsBaseType(metaclass=VisionsBaseTypeMeta):
             A new type
         """
         new_type = type(
-                    "{name}[{type_name}]".format(name=cls.__name__, type_name=type_name),
-                    (cls,),
-                    {
-                        "get_relations": classmethod(lambda _: NotImplemented),
-                        "contains_op": cls.contains_op,
-                    },
+            "{name}[{type_name}]".format(name=cls.__name__, type_name=type_name),
+            (cls,),
+            {
+                "get_relations": classmethod(lambda _: NotImplemented),
+                "contains_op": cls.contains_op,
+            },
         )
-        new_relations = list(relations_generator(new_type)) if relations_generator else []
+        new_relations = (
+            list(relations_generator(new_type)) if relations_generator else []
+        )
         if replace:
-            assert relations_generator is not None, "When calling evolve_type with `replace=True`, a `relations_generator` is required."
+            assert (
+                relations_generator is not None
+            ), "When calling evolve_type with `replace=True`, a `relations_generator` is required."
             relations_method = classmethod(lambda _: new_relations)
         else:
             old_relations = [attr.evolve(x, type=new_type) for x in cls.get_relations()]
