### 2020-03-11 dylan-profiler/visions#22 @1313e
Correct, that is another way to do it.
Currently however, the class is correctly named `Float` and therefore does not clash with the built-in `float` type (but the module in which this class lives, does obviously).
Also, if someone executes `from visions.types import float`, they kind of made that mistake by themselves.
It is not really the job of the programmer to avoid potential namespace clashes that are extremely easy to avoid by the user itself.
Common Python coding practice is also to never import any definition from a namespace, but instead import the namespace itself (in this case `visions`) and access all definitions from there.

Keep however in mind that not putting the modules in `__all__` does not mean that they aren't imported and therefore accessible by the user.
The `__all__` dunder attribute is solely used for determining what definitions should be imported when someone performs a wildcard import.
It therefore is important to have one to avoid importing more than is necessary, but the `float` module will still be available as `visions.types.float` regardless of it being in any `__all__` declaration.
