### 2020-03-11 dylan-profiler/visions#31 @sbrugman
To be able to use the `CompleteTypeset`, there are additional dependencies:

```pip install visions[type_geometry]```
(more information https://dylan-profiler.github.io/visions/visions/getting_started/installation.html)

The loading of the requirements could be more efficient. The pending PR #27 contains some improvements already.
