### 2020-03-11 dylan-profiler/visions#31 @1313e
The example in the documentation does not mention that the additional dependencies are required for that example.
Also, you may want to avoid having several optional additional dependencies for your package, as that is very confusing for average users.
As long as a package is not incredibly big (like `PyQt5`) or hard to install (like `mpi4py`), you could opt for simply installing it by default as long as it is not solely used for testing.
