### 2020-03-11 0ccf97d
reworked for mypy
