+++ src/visions/types/type.py
@@ -18,7 +18,7 @@ class VisionsBaseTypeMeta(ABCMeta):
 
     @property
     def relations(cls):
-        if clas._relations is None:
+        if cls._relations is None:
             cls._relations = cls.get_relations()
         return cls._relations
 
@@ -58,7 +58,9 @@ class VisionsBaseType(metaclass=VisionsBaseTypeMeta):
     def evolve_type(
         cls,
         type_name: str,
-        relations_generator: Optional[Callable] = None,
+        relations_generator: Optional[
+            Callable[[Type[VisionsBaseTypeMeta]], Sequence[TypeRelation]]
+        ] = None,
         replace: bool = False,
     ):
         """Make a copy of the type with the relations replaced by the relations return by `relations_func`.
@@ -70,13 +72,14 @@ class VisionsBaseType(metaclass=VisionsBaseTypeMeta):
         Returns:
             A new type
         """
+
+        def get_new_relations():
+            return relations
+
         new_type = type(
             "{name}[{type_name}]".format(name=cls.__name__, type_name=type_name),
             (cls,),
-            {
-                "get_relations": classmethod(lambda _: NotImplemented),
-                "contains_op": cls.contains_op,
-            },
+            {"get_relations": get_new_relations, "contains_op": cls.contains_op},
         )
         new_relations = (
             list(relations_generator(new_type)) if relations_generator else []
@@ -85,9 +88,9 @@ class VisionsBaseType(metaclass=VisionsBaseTypeMeta):
             assert (
                 relations_generator is not None
             ), "When calling evolve_type with `replace=True`, a `relations_generator` is required."
-            relations_method = classmethod(lambda _: new_relations)
+            relations = new_relations
         else:
             old_relations = [attr.evolve(x, type=new_type) for x in cls.get_relations()]
-            relations_method = classmethod(lambda _: old_relations + new_relations)
-        new_type.get_relations = relations_method
+            relations = old_relations + new_relations
+
         return new_type
