### 2020-03-11 edd2db4
more mypy fixes
