+++ mypy.ini
@@ -29,3 +29,6 @@ ignore_missing_imports = True
 
 [mypy-matplotlib.*]
 ignore_missing_imports = True
+
+[mypy-tangled_up_in_unicode.*]
+ignore_missing_imports = True
