### 2020-03-11 8241d1f
- Functional API (#22) - Test replace xfail with assertions (#32) - `__all__` dunder attribute for types and typesets - Ordering of imports (standard lib, external packages, same package) - Drop `pytest-ordering` requirement
