+++ src/visions/application/summaries/series/text_summary.py
@@ -1,5 +1,6 @@
 from collections import Counter
 from typing import Dict
+
 import pandas as pd
 
 from tangled_up_in_unicode import category, category_long, script, block, block_abbr
