+++ src/visions/types/complex.py
@@ -1,7 +1,8 @@
+from typing import Sequence
+
 import pandas.api.types as pdt
 import pandas as pd
 import numpy as np
-from typing import Sequence
 
 from visions.types.float import test_string_is_float
 from visions.relations import IdentityRelation, InferenceRelation, TypeRelation
@@ -11,7 +12,6 @@ from visions.utils.coercion import test_utils
 
 def test_string_is_complex(series) -> bool:
     coerced_series = test_utils.option_coercion_evaluator(to_complex)(series)
-
     return coerced_series is not None and not test_string_is_float(series)
 
 
