+++ src/visions/types/path.py
@@ -1,5 +1,6 @@
-import pathlib
 from typing import Sequence
+import pathlib
+
 import pandas as pd
 
 from visions.relations import IdentityRelation, InferenceRelation, TypeRelation
