+++ src/visions/types/categorical.py
@@ -1,6 +1,7 @@
+from typing import Sequence
+
 import pandas.api.types as pdt
 import pandas as pd
-from typing import Sequence, Type
 
 from visions.relations import IdentityRelation, TypeRelation
 from visions.types import VisionsBaseType
