+++ tests/test_type_evolve.py
@@ -3,6 +3,7 @@ from visions.relations import IdentityRelation, TypeRelation
 from visions.relations.integer_to_datetime import integer_to_datetime_year_month_day
 from visions.types.integer import Integer
 
+
 def relations_generator(vtype):
     return [integer_to_datetime_year_month_day(vtype)]
 
@@ -16,7 +17,9 @@ def test_evolve_extend_datetime():
 
 
 def test_evolve_replace_datetime():
-    DateTimeIntYYYYMMDD = DateTime.evolve_type("int_yyyymmdd_r", relations_generator, replace=True)
+    DateTimeIntYYYYMMDD = DateTime.evolve_type(
+        "int_yyyymmdd_r", relations_generator, replace=True
+    )
     assert str(DateTimeIntYYYYMMDD) == "DateTime[int_yyyymmdd_r]"
     assert str(DateTime) == "DateTime"
     assert len(DateTime.get_relations()) == 2
