+++ src/visions/types/type.py
@@ -62,12 +62,12 @@ class VisionsBaseType(metaclass=VisionsBaseTypeMeta):
         ] = None,
         replace: bool = False,
     ):
-        """Make a copy of the type with the relations replaced by the relations return by `relations_func`.
+        """Make a copy of the type
 
         Args:
             type_name: the new type suffix, the type name will be `type[type_name]`
             relations_generator: a function returning all TypeRelations for the new type
-            replace: ...
+            replace: if True, do not include the existing relations
 
         Returns:
             A new type
@@ -90,7 +90,7 @@ class VisionsBaseType(metaclass=VisionsBaseTypeMeta):
             ), "When calling evolve_type with `replace=True`, a `relations_generator` is required."
             relations = new_relations
         else:
-            old_relations = [attr.evolve(x, type=new_type) for x in cls.get_relations()]
+            old_relations = [attr.evolve(relation, type=new_type) for relation in cls.get_relations()]
             relations = old_relations + new_relations
 
         return new_type
