### 2020-03-11 5f30030
Rebased #19 on master without docs build (#27)
- Clarify the purpose of the application section (#23).
- Type evolving (and tests)
- Functional API (#22)
- Test replace xfail with assertions (#32)
- `__all__` dunder attribute for types and typesets
- Ordering of imports (standard lib, external packages, same package)
- Drop `pytest-ordering` requirement
- Add a support page
- Fix typing issues

Co-authored-by: Ian Eaves <ian.k.eaves@gmail.com>
