+++ src/visions/relations/integer_to_datetime.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from visions.types import DateTime
+from visions.types import Integer
 from visions.relations import InferenceRelation
 from visions.relations.string_to_datetime import to_datetime_year_month_day
 from visions.utils.coercion import test_utils
@@ -10,22 +10,19 @@ def to_datetime(series: pd.Series) -> pd.Series:
     return pd.to_datetime(series)
 
 
+def _to_datetime(cls, func) -> InferenceRelation:
+    return InferenceRelation(
+        relationship=test_utils.coercion_test(lambda s: func(s.astype(str))),
+        transformer=to_datetime,
+        type=cls,
+        related_type=Integer,
+    )
+
+
 # TODO: do only convert obvious dates (20191003000000)
 def integer_to_datetime(cls):
-    return InferenceRelation(
-        relationship=test_utils.coercion_test(lambda s: to_datetime(s.astype(str))),
-        transformer=to_datetime,
-        type=DateTime,
-        related_type=cls,
-    )
+    return _to_datetime(cls, to_datetime)
 
 
-def integer_to_datetime_year_month_day(cls):
-    return InferenceRelation(
-        relationship=test_utils.coercion_test(
-            lambda s: to_datetime_year_month_day(s.astype(str))
-        ),
-        transformer=to_datetime,
-        type=DateTime,
-        related_type=cls,
-    )
+def integer_to_datetime_year_month_day(cls) -> InferenceRelation:
+    return _to_datetime(cls, to_datetime_year_month_day)
