+++ src/visions/application/summaries/series/text_summary.py
@@ -1,8 +1,21 @@
 from collections import Counter
+from typing import Dict
 
 import pandas as pd
 
 from tangled_up_in_unicode import category, category_long, script, block, block_abbr
+
+
+def get_character_counts(series) -> Dict[str, int]:
+    """Function to return the character counts
+
+    Args:
+        series: the Series to process
+
+    Returns:
+        A dict with character counts
+    """
+    return dict(Counter(series.str.cat()))
 
 
 def text_summary(series: pd.Series) -> dict:
@@ -18,7 +31,7 @@ def text_summary(series: pd.Series) -> dict:
     summary = {"length": series.map(lambda x: len(str(x))).value_counts().to_dict()}
 
     # Unicode Character Summaries (category and script name)
-    character_counts = dict(Counter(series.str.cat()))
+    character_counts = get_character_counts(series)
 
     summary["category_short_values"] = {
         key: category(key) for key in character_counts.keys()
