+++ src/visions/types/generic.py
@@ -1,5 +1,6 @@
-import pandas as pd
 from typing import Sequence
+
+import pandas as pd
 
 from visions.types import VisionsBaseType
 from visions.relations import TypeRelation
