+++ src/visions/types/uuid.py
@@ -1,7 +1,7 @@
 import uuid
+from typing import Sequence
 
 import pandas as pd
-from typing import Sequence
 
 from visions.relations import IdentityRelation, InferenceRelation, TypeRelation
 from visions.types.type import VisionsBaseType
