+++ docs/_sources/visions/api/_autosummary/visions.functional.infer_series_type.rst.txt
@@ -1,0 +1,6 @@
+visions.functional.infer\_series\_type
+======================================
+
+.. currentmodule:: visions.functional
+
+.. autofunction:: infer_series_type
