+++ docs/_sources/visions/api/_autosummary/visions.functional.infer_frame_type.rst.txt
@@ -1,0 +1,6 @@
+visions.functional.infer\_frame\_type
+=====================================
+
+.. currentmodule:: visions.functional
+
+.. autofunction:: infer_frame_type
