+++ docs/_sources/visions/api/_autosummary/visions.functional.detect_frame_type.rst.txt
@@ -1,0 +1,6 @@
+visions.functional.detect\_frame\_type
+======================================
+
+.. currentmodule:: visions.functional
+
+.. autofunction:: detect_frame_type
