+++ docs/_sources/visions/api/_autosummary/visions.functional.cast_series.rst.txt
@@ -1,0 +1,6 @@
+visions.functional.cast\_series
+===============================
+
+.. currentmodule:: visions.functional
+
+.. autofunction:: cast_series
