+++ docs/_sources/visions/api/_autosummary/visions.functional.cast_frame.rst.txt
@@ -1,0 +1,6 @@
+visions.functional.cast\_frame
+==============================
+
+.. currentmodule:: visions.functional
+
+.. autofunction:: cast_frame
