### 2020-03-11 dylan-profiler/visions#30 @sbrugman
Agreed. There is now a dedicated support page:
https://dylan-profiler.github.io/visions/visions/contributing/support.html
