+++ src/visions/typesets/geometry_set.py
@@ -45,3 +45,13 @@ class GeometrySet(VisionsTypeset):
             Geometry,
         }
         super().__init__(types)
+
+        try:
+            import shapely
+        except ImportError as e:
+            raise ImportError(
+                "This typeset requires dependencies that are currently not installed ({}). "
+                "You can follow the installation instructions to resolve this issue: "
+                "https://dylan-profiler.github.io/visions/visions/getting_started/installation.html"
+                "".format(str(e))
+            )
