+++ src/visions/__init__.py
@@ -6,6 +6,8 @@ from visions.dtypes.boolean import BoolDtype
 from visions import types, typesets
 
 from visions.types import *
+from visions.typesets import *
+
 from visions.functional import (
     cast_frame,
     infer_frame_type,
