+++ docs/_sources/visions/api/_autosummary/visions.functional.compare_detect_inference_frame.rst.txt
@@ -1,0 +1,6 @@
+visions.functional.compare\_detect\_inference\_frame
+====================================================
+
+.. currentmodule:: visions.functional
+
+.. autofunction:: compare_detect_inference_frame
