+++ docs/_sources/visions/api/_autosummary/visions.functional.type_inference_frame.rst.txt
@@ -1,6 +1,0 @@
-visions.functional.type\_inference\_frame
-=========================================
-
-.. currentmodule:: visions.functional
-
-.. autofunction:: type_inference_frame
