+++ docs/_sources/visions/api/_autosummary/visions.functional.type_cast_frame.rst.txt
@@ -1,6 +1,0 @@
-visions.functional.type\_cast\_frame
-====================================
-
-.. currentmodule:: visions.functional
-
-.. autofunction:: type_cast_frame
