+++ docs/_sources/visions/api/_autosummary/visions.functional.type_cast_series.rst.txt
@@ -1,6 +1,0 @@
-visions.functional.type\_cast\_series
-=====================================
-
-.. currentmodule:: visions.functional
-
-.. autofunction:: type_cast_series
