### 2020-03-11 dylan-profiler/visions#31 @sbrugman
The default installation instructions include all dependencies (```pip install visions[all]```), for which all examples work. We chose to make `visions` lightweight by default as it is included in the dependency tree of a frequently downloaded package.

For future reference, we are raising an exception when instantiating a typeset when dependencies are missing.

> ValueError: This typeset requires dependencies that are currently not installed (No module named 'shapely'). You can follow the installation instructions to resolve this issue: https://dylan-profiler.github.io/visions/visions/getting_started/installation.html
