+++ requirements.txt
@@ -1,5 +1,5 @@
 numpy
-pandas==0.25.3
+pandas==1.0.1
 attrs==19.3.0
 networkx==2.4
 tangled_up_in_unicode==0.0.4
