### 2020-03-11 9a5b0f3
Update MANIFEST.in
