+++ MANIFEST.in
@@ -1,2 +1,4 @@
 include LICENSE
 include README.md
+include requirements.txt
+include requirements_test.txt
