### 2020-03-12 #2145 @mjsottile
The only thing I found when reviewing the paper and repository was that the first word of Figure 1's caption should be capitalized.

Given that everyone seems to agree that the paper is good to go forward, the next steps for @sbrugman are:

- Submit a version of the software to an archive (e.g., figshare or zenodo) and post the DOI for the archive in this thread.

- Please verify that the version number (0.3.0) matches that of the archived code, and if not, let me know so I can update the version here.

- Finally, make sure that the metadata for the archive (e.g., title, authors) matches the paper.

Once all of these steps are done, I'll pass the paper on up the line to the EIC team for finalizing the process.
