+++ tests/test_type_property.py
@@ -9,5 +9,9 @@ def test_property_base():
 
 def test_property_types():
     """Assert that the relations property contains a list"""
-    assert type(Integer.relations) == list and all(isinstance(r, TypeRelation) for r in Integer.relations)
-    assert type(IPAddress.relations) == list and all(isinstance(r, TypeRelation) for r in IPAddress.relations)
+    assert type(Integer.relations) == list and all(
+        isinstance(r, TypeRelation) for r in Integer.relations
+    )
+    assert type(IPAddress.relations) == list and all(
+        isinstance(r, TypeRelation) for r in IPAddress.relations
+    )
