### 2020-04-02 ae89c85
Removed code that does not belong to this feature
