+++ .travis.yml
@@ -3,7 +3,6 @@ dist: bionic
 language: python
 cache: pip
 python:
-  - "3.5"
   - "3.6"
   - "3.7"
   - "3.8"
