### 2020-04-06 35eee5e
Robustify PyPi deployment (check manifest and distribution)
