### 2020-04-06 82519ef
Robustify PyPi deployment (check manifest and distribution)
