+++ .travis.yml
@@ -16,6 +16,7 @@ install:
   - check-manifest
   - python setup.py sdist bdist_wheel
   - twine check dist/*
+  - pip install -e .
 
 script:
 #  - pytest --cov=. tests/
