### 2020-04-06 f4179c4
Separate changes that can be merged now from those that need some more work.
Merged now:
- Integer should be a strict subset of Float
- Fix pandas future warning
