### 2020-04-06 9e550d0
Integer as subset of Float
Changes: 

- Integer should be a strict subset of Float
- Fix pandas future warning
