+++ .travis.yml
@@ -2,10 +2,15 @@ sudo: required
 dist: bionic
 language: python
 cache: pip
-python:
-  - "3.6"
-  - "3.7"
-  - "3.8"
+
+jobs:
+  include:
+    - python: 3.6
+    - python: 3.7
+    - python: 3.8
+    - python: 3.9-dev
+  allow_failures:
+    - python: 3.9-dev
 
 before_install:
   - python -m pip install --upgrade pip setuptools wheel
