### 2020-04-06 1ad809a
Extended testing for Pypi release (#43)
