+++ setup.py
@@ -42,7 +42,7 @@ extras_requires["all"] = requirements + [
 
 setup(
     name="visions",
-    version="0.4.0",
+    version="0.4.1",
     url="https://github.com/dylan-profiler/visions",
     description="Visions",
     author="Dylan Profiler",
