+++ paper/paper.md
@@ -79,7 +79,7 @@ The package aims to make defining and relating data types simple while maximizin
 This enables users to define, relate, modify and share types to solve data tasks and has direct applications to exploratory data analytics (EDA) problem such as those found in [@pandasprofiling] where the appropriate statistics for data of different types varies greatly.
 It further paves the way for exciting future opportunities to identify and understand semantically useful data abstractions across a variety of potential domains. 
 
-`visions` builds on `pandas` [@mckinney2010data] and `numpy` [@van2011numpy] for machine type representations. 
+`visions` builds on `pandas` [@mckinney2010data; @reback2020pandas] and `numpy` [@van2011numpy] for machine type representations. 
 Other work has created specific semantic data types for pandas, namely `cyberpandas` ([1], [2], [3]) and `geopandas` [@kelsey_jordahl_2019_3545747].
 Moreover, `networkx` [@SciPyProceedings_11] is employed to construct type relation graphs.
 
