+++ paper/paper.bib
@@ -9,6 +9,17 @@
   organization={Austin, TX}
 }
 
+ @software{reback2020pandas,
+    author       = {The pandas development team},
+    title        = {pandas-dev/pandas: Pandas 1.0.1},
+    month        = feb,
+    year         = 2020,
+    publisher    = {Zenodo},
+    version      = {v1.0.1},
+    doi          = {10.5281/zenodo.3644238},
+    url          = {https://doi.org/10.5281/zenodo.3644238}
+}
+
 # Numpy
 @article{van2011numpy,
   journal={Computing in Science Engineering},
