### 2020-04-11 7dbce3f
Catch edge case: series with exclusively missing values
