+++ tests/test_type_convert.py
@@ -120,13 +120,13 @@ def test_consistency(series):
         converted_series = typeset.cast_series(series.copy(deep=True))
         assert series.dtype.kind != converted_series.dtype.kind or not (
             (
-                converted_series.eq(series) ^ (converted_series.isna() & series.isna())
+                converted_series.eq(series) | (converted_series.isna() & series.isna())
             ).all()
         )
     else:
         converted_series = typeset.cast_series(series.copy(deep=True))
         assert (
-            converted_series.eq(series) ^ (converted_series.isna() & series.isna())
+            converted_series.eq(series) | (converted_series.isna() & series.isna())
         ).all()
 
 
