### 2020-04-11 #2145 @sbrugman
We went through the remaining steps.
- The software version has increased.  For the publication is v0.4.1.
- For the archive, the Zenodo DOI is http://doi.org/10.5281/zenodo.3748758.

@1313e and @fccoelho Many thanks for your constructive reviews!

@mjsottile Thanks for editing this one :)
