### 2020-04-11 #2145 @arfon
@sbrugman - a couple of quick changes please for us to proceed from here. Could you please add `10.25080/majora-92bf1922-00a` to the `doi` field of your BibTeX entry for that paper Whedon is suggesting (it looks to be a correct DOI suggestion).

Also, the fact that you have a `*` after your author names is causing issues with our production pipeline. Could you explain what you're trying to accomplish here and perhaps I can make an alternative suggestion?
