+++ paper/paper.bib
@@ -6,7 +6,8 @@
   volume={445},
   pages={51--56},
   year={2010},
-  organization={Austin, TX}
+  organization={Austin, TX},
+  doi={10.25080/majora-92bf1922-00a}
 }
 
  @software{reback2020pandas,
