+++ paper/paper.bib
@@ -1,13 +1,12 @@
 # Pandas
-@inproceedings{mckinney2010data,
-  title={{D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython},
-  author={McKinney, Wes and others},
-  booktitle={Proceedings of the 9th Python in Science Conference},
-  volume={445},
-  pages={51--56},
-  year={2010},
-  organization={Austin, TX},
-  doi={10.25080/majora-92bf1922-00a}
+@InProceedings{ mckinney-proc-scipy-2010,
+  author    = { {W}es {M}c{K}inney },
+  title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
+  booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
+  pages     = { 56 - 61 },
+  year      = { 2010 },
+  editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
+  doi       = { 10.25080/Majora-92bf1922-00a }
 }
 
  @software{reback2020pandas,
