### 2020-04-11 #2145 @sbrugman
> Also, the fact that you have a * after your author names is causing issues with our production pipeline. Could you explain what you're trying to accomplish here and perhaps I can make an alternative suggestion?

The * is intended to indicate equal contribution.
