### 2020-04-11 #2145 @arfon
@sbrugman -  could you merge this PR (https://github.com/dylan-profiler/visions/pull/48) or make a similar change?
