### 2020-04-11 #2145 @arfon
Can you check the new pdf proof looks ok (for denoting equal authorship?)
